+++ crates/feagi-brain-development/src/connectivity/core_morphologies/vectors.rs
@@ -20,8 +20,8 @@ pub fn apply_vectors_morphology_with_dimensions(
     dst_area_id: u32,
     vectors: Vec<(i32, i32, i32)>,
     dst_dimensions: (usize, usize, usize),
-    weight: u8,
-    psp: u8,
+    weight: f32,
+    psp: f32,
     synapse_attractivity: u8,
     synapse_type: SynapseType,
 ) -> BduResult<u32> {
@@ -101,8 +101,8 @@ pub fn apply_vectors_morphology(
     src_area_id: u32,
     dst_area_id: u32,
     vectors: Vec<(i32, i32, i32)>,
-    weight: u8,
-    psp: u8,
+    weight: f32,
+    psp: f32,
     synapse_attractivity: u8,
     synapse_type: SynapseType,
 ) -> BduResult<u32> {
